feat : 3단계 기능 구현

+++ src/main/java/roomescape/Reservation.java
@@ -5,19 +5,19 @@ import java.time.LocalTime;
 
 public class Reservation {
 
-    private final int id;
+    private final long id;
     private final String name;
     private final String date;
     private final String time;
 
-    public Reservation(int id, String name, String date, String time) {
+    public Reservation(long id, String name, String date, String time) {
         this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
